Move application and experiment generation code to internal/
