+++ internal/application/experiment/application.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 
 	redskyv1beta1 "github.com/thestormforge/optimize-controller/api/v1beta1"
-	"github.com/thestormforge/optimize-controller/redskyctl/internal/commands/generate/experiment/k8s"
+	"github.com/thestormforge/optimize-controller/internal/application/experiment/k8s"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/kustomize/kyaml/yaml"
 )
