+++ internal/application/experiment/generator.go
@@ -21,8 +21,8 @@ import (
 
 	redskyappsv1alpha1 "github.com/thestormforge/optimize-controller/api/apps/v1alpha1"
 	redskyv1beta1 "github.com/thestormforge/optimize-controller/api/v1beta1"
+	"github.com/thestormforge/optimize-controller/internal/application"
 	meta2 "github.com/thestormforge/optimize-controller/internal/meta"
-	"github.com/thestormforge/optimize-controller/pkg/application"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
