+++ internal/application/experiment/locust/scenarios.go
@@ -20,8 +20,8 @@ import (
 	"fmt"
 
 	redskyappsv1alpha1 "github.com/thestormforge/optimize-controller/api/apps/v1alpha1"
-	"github.com/thestormforge/optimize-controller/redskyctl/internal/commands/generate/experiment/k8s"
-	"github.com/thestormforge/optimize-controller/redskyctl/internal/commands/generate/experiment/prometheus"
+	"github.com/thestormforge/optimize-controller/internal/application/experiment/k8s"
+	"github.com/thestormforge/optimize-controller/internal/application/experiment/prometheus"
 	batchv1beta1 "k8s.io/api/batch/v1beta1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
