+++ internal/application/experiment/objectives.go
@@ -22,8 +22,8 @@ import (
 	"strings"
 
 	redskyappsv1alpha1 "github.com/thestormforge/optimize-controller/api/apps/v1alpha1"
-	"github.com/thestormforge/optimize-controller/redskyctl/internal/commands/generate/experiment/k8s"
-	"github.com/thestormforge/optimize-controller/redskyctl/internal/commands/generate/experiment/prometheus"
+	"github.com/thestormforge/optimize-controller/internal/application/experiment/k8s"
+	"github.com/thestormforge/optimize-controller/internal/application/experiment/prometheus"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 )
