+++ internal/application/experiment/prometheus/setuptask.go
@@ -18,8 +18,8 @@ package prometheus
 
 import (
 	redskyv1beta1 "github.com/thestormforge/optimize-controller/api/v1beta1"
+	"github.com/thestormforge/optimize-controller/internal/application/experiment/k8s"
 	"github.com/thestormforge/optimize-controller/internal/setup"
-	"github.com/thestormforge/optimize-controller/redskyctl/internal/commands/generate/experiment/k8s"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
