+++ redskyctl/internal/commands/export/export.go
@@ -29,13 +29,13 @@ import (
 	"github.com/spf13/cobra"
 	app "github.com/thestormforge/optimize-controller/api/apps/v1alpha1"
 	redsky "github.com/thestormforge/optimize-controller/api/v1beta1"
+	apppkg "github.com/thestormforge/optimize-controller/internal/application"
+	experimentctl "github.com/thestormforge/optimize-controller/internal/application/experiment"
 	"github.com/thestormforge/optimize-controller/internal/experiment"
 	"github.com/thestormforge/optimize-controller/internal/patch"
 	"github.com/thestormforge/optimize-controller/internal/server"
 	"github.com/thestormforge/optimize-controller/internal/template"
-	apppkg "github.com/thestormforge/optimize-controller/pkg/application"
 	"github.com/thestormforge/optimize-controller/redskyctl/internal/commander"
-	experimentctl "github.com/thestormforge/optimize-controller/redskyctl/internal/commands/generate/experiment"
 	"github.com/thestormforge/optimize-controller/redskyctl/internal/kustomize"
 	experimentsapi "github.com/thestormforge/optimize-go/pkg/api/experiments/v1alpha1"
 	"github.com/thestormforge/optimize-go/pkg/config"
